requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
python-frontmatter>=1.1
pypdf>=5.0
pdfminer.six>=20231228
httpx>=0.28
websockets>=12
watchdog>=6.0
pyyaml>=6.0
pdf2image>=1.17
arxiv>=2.0.0
